requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
distlib==0.3.7
Django==4.2.8
django-filter==23.5
djangorestframework==3.14.0
filelock==3.13.1
importlib-metadata==7.0.0
Markdown==3.5.1
pbr==6.0.0
platformdirs==3.11.0
pytz==2023.3.post1
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.9.0
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
zipp==3.17.0
